bhjgbhkjhkjhk

+++ code/port_mapper/Includes/mapper.h
@@ -27,7 +27,7 @@ typedef struct port_mapper{
 
 	
 	int32_t* port_arr; //its a 32 bit int because it makes using my mtx functions more convenient
-	int32_t curr_num_ports;
+	uint16_t curr_num_ports;
 	int socket;
 	struct sockaddr_in addr_struct;
 	fd_set con_fds;
